# be/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PlanWise API Documentation",
        "description": "API Documentation for PlanWise application - a smart daily planning application with AI features.",
        "contact": {
            "name": "PlanWise Team",
            "email": "contact@planwise.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://springdoc.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development"
        }
    ],
    "security": [
        {
            "bearer": []
        }
    ],
    "paths": {
...
        "/api/v1/tasks/{taskId}": {
            "get": {
                "tags": [
                    "task-controller"
                ],
                "operationId": "getTaskById",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "task-controller"
                ],
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "task-controller"
                ],
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "task-controller"
                ],
                "operationId": "getTasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "task-controller"
                ],
                "operationId": "createTask",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTaskRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/subscriptions/verify": {
            "post": {
                "tags": [
                    "subscription-controller"
                ],
                "operationId": "verifyPayment",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MomoIPNRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/subscriptions/purchase": {
            "post": {
                "tags": [
                    "subscription-controller"
                ],
                "operationId": "purchasePlan",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/subscriptions/momo-ipn": {
            "post": {
                "tags": [
                    "subscription-controller"
                ],
                "operationId": "receiveMomoIPN",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MomoIPNRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/subscriptions/mock-ipn": {
            "post": {
                "tags": [
                    "subscription-controller"
                ],
                "operationId": "mockIpn",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/subscriptions/vnpay-return": {
            "get": {
                "tags": [
                    "subscription-controller"
                ],
                "operationId": "vnpayReturn",
                "parameters": [
                    {
                        "name": "vnpParams",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/subscriptions/vnpay-ipn": {
            "get": {
                "tags": [
                    "subscription-controller"
                ],
                "operationId": "vnpayIpn",
                "parameters": [
                    {
                        "name": "vnpParams",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/subscriptions/transactions/{orderId}/status": {
            "get": {
                "tags": [
                    "subscription-controller"
                ],
                "operationId": "getTransactionStatus",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateTaskRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "dueDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    },
                    "priority": {
                        "type": "string",
                        "example": "MEDIUM"
                    },
                    "color": {
                        "type": "string",
                        "example": "indigo"
                    },
                    "categoryId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "eisenhowerMatrix": {
                        "type": "string",
                        "example": "urgent_important"
                    },
                    "estimatedTime": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "contexts": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "sortOrder": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "TaskDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "dueDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    },
                    "priority": {
                        "type": "string",
                        "example": "MEDIUM"
                    },
                    "color": {
                        "type": "string",
                        "example": "indigo"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "eisenhowerMatrix": {
                        "type": "string",
                        "example": "urgent_important"
                    },
                    "estimatedTime": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "actualTime": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "contexts": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "categoryId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "categoryName": {
                        "type": "string"
                    },
                    "categoryColor": {
                        "type": "string"
                    },
                    "sortOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    }
                }
            },
...
            "CreateTaskRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "dueDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    },
                    "priority": {
                        "type": "string",
                        "example": "MEDIUM"
                    },
                    "color": {
                        "type": "string",
                        "example": "indigo"
                    },
                    "categoryId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "eisenhowerMatrix": {
                        "type": "string",
                        "example": "urgent_important"
                    },
                    "estimatedTime": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "contexts": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "title"
                ]
            },
            "MomoIPNRequest": {
                "type": "object",
                "properties": {
                    "partnerCode": {
                        "type": "string"
                    },
                    "orderId": {
                        "type": "string"
                    },
                    "requestId": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "orderInfo": {
                        "type": "string"
                    },
                    "orderType": {
                        "type": "string"
                    },
                    "transId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "resultCode": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "payType": {
                        "type": "string"
                    },
                    "responseTime": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "extraData": {
                        "type": "string"
                    },
                    "signature": {
                        "type": "string"
                    }
                }
            },
...
            "TaskListResponse": {
                "type": "object",
                "properties": {
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskDto"
                        }
                    },
                    "totalCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pendingCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "completedCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "overdueCount": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter your Bearer JWT token to access protected endpoints.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}